@@ -1,4 +1,4 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.2
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "FirebaseAPI",
     platforms: [
-        .iOS(.v15), .macOS(.v10_15)
+        .macOS(.v15), .iOS(.v18), .watchOS(.v11), .tvOS(.v18), .visionOS(.v2)
     ],
     products: [
         .library(
@@ -14,23 +14,26 @@ let package = Package(
             targets: ["FirestoreAPI"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-protobuf.git", from: "1.25.2"),
-        .package(url: "https://github.com/grpc/grpc-swift.git", from: "1.24.1"),
-        .package(url: "https://github.com/apple/swift-log.git", branch: "main")
+        .package(url: "https://github.com/apple/swift-protobuf.git", from: "1.31.1"),
+        .package(url: "https://github.com/grpc/grpc-swift-2.git", from: "2.1.0"),
+        .package(url: "https://github.com/grpc/grpc-swift-protobuf.git", from: "2.0.0"),
+        .package(url: "https://github.com/apple/swift-log.git", from: "1.6.4")
     ],
     targets: [
         .target(
             name: "FirestoreAPI",
             dependencies: [
                 .product(name: "Logging", package: "swift-log"),
-                .product(name: "GRPC", package: "grpc-swift"),
-                .product(name: "protoc-gen-swift", package: "swift-protobuf")
+                .product(name: "GRPCCore", package: "grpc-swift-2"),
+                .product(name: "GRPCProtobuf", package: "grpc-swift-protobuf"),
+                .product(name: "SwiftProtobuf", package: "swift-protobuf")
             ]),
         .testTarget(
             name: "FirebaseAPITests",
             dependencies: [
-                .product(name: "GRPC", package: "grpc-swift"),
-                .product(name: "protoc-gen-swift", package: "swift-protobuf"),
+                .product(name: "GRPCCore", package: "grpc-swift-2"),
+                .product(name: "GRPCProtobuf", package: "grpc-swift-protobuf"),
+                .product(name: "SwiftProtobuf", package: "swift-protobuf"),
                 "FirestoreAPI"
             ]),
     ]
